refactor: move sql util
